Fix null reference warnings when upgrading to .NET 6
In .NET6 templates, nullable reference types are enabled by default.
https://github.com/edc-it/MEInsight/blob/61a80d3ed6a4ffb5dd2396c378febbcc660bdb37/MEInsight.Web/MEInsight.Web.csproj#L4-L5

https://docs.microsoft.com/en-us/dotnet/csharp/nullable-references

These updates resolves nullable warnings. Testing is pending.

#8

+++ MEInsight.Web/Areas/Settings/Controllers/AssessmentTypesController.cs
@@ -150,7 +150,7 @@ namespace MEInsight.Web.Areas.Settings.Controllers
 
             int relatedCount = 0;
                 
-            relatedCount += refAssessmentType.ProgramAssessments.Count();
+            relatedCount += refAssessmentType.ProgramAssessments.Count;
 
             if(relatedCount > 0)
             {
@@ -171,9 +171,14 @@ namespace MEInsight.Web.Areas.Settings.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+
             var refAssessmentType = await _context.AssessmentTypes.FindAsync(id);
 
-            _context.AssessmentTypes.Remove(refAssessmentType);
+            if (refAssessmentType != null)
+            {
+                _context.AssessmentTypes.Remove(refAssessmentType);
+            }
+            
             await _context.SaveChangesAsync();
         
             TempData["messageType"] = "success";
