+++ MEInsight.Web/Controllers/ProgramsController.cs
@@ -174,7 +174,8 @@ namespace MEL.Web.Controllers
         [Authorize(Policy = "RequireDeleteRole")]
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+
+            if (id == null || _context.Programs == null)
             {
                 return NotFound();
             }
@@ -215,9 +216,18 @@ namespace MEL.Web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+
+            if (_context.Programs == null)
+            {
+                return Problem("Entity set 'ApplicationDbContext.Program'  is null.");
+            }
             var program = await _context.Programs.FindAsync(id);
 
-            _context.Programs.Remove(program);
+            if (program != null)
+            {
+                _context.Programs.Remove(program);
+            }
+                        
             await _context.SaveChangesAsync();
         
             TempData["messageType"] = "success";
